requirements.txt
Flask==0.12.1
azure-storage-blob
scikit-learn
scipy
